Add `"state"` to `authn.models.Authenticator` (PAN-15469)
Also added other missing properties. Many of these do not have
descriptions in the API documentation so the docs here are skim.

`isEnable()` has been renamed to `isEnabled()` to match the API's
response schema. This is a breaking change but the current property
isn't being deserialized properly anyways so it's been unusable this
whole time.